+++ sphinx_needs/directives/needbar.py
@@ -52,6 +52,8 @@ class NeedbarDirective(FilterBase):
         "legend": directives.flag,
         "stacked": directives.flag,
         "show_sum": directives.flag,
+        "show_top_sum": directives.flag,
+        "sum_rotation": directives.unchanged_required,
         "transpose": directives.flag,
         "horizontal": directives.flag,
     }
@@ -119,6 +121,10 @@ class NeedbarDirective(FilterBase):
 
         stacked = "stacked" in self.options
         show_sum = "show_sum" in self.options
+        show_top_sum = "show_top_sum" in self.options
+        sum_rotation = self.options.get("sum_rotation")
+        if sum_rotation:
+            sum_rotation = sum_rotation.strip()
         transpose = "transpose" in self.options
         horizontal = "horizontal" in self.options
 
@@ -141,6 +147,8 @@ class NeedbarDirective(FilterBase):
             "separator": separator,
             "stacked": stacked,
             "show_sum": show_sum,
+            "show_top_sum": show_top_sum,
+            "sum_rotation": sum_rotation,
             "transpose": transpose,
             "horizontal": horizontal,
             "style": style,
@@ -324,6 +332,7 @@ def process_needbar(app: Sphinx, doctree: nodes.document, fromdocname: str):
         y_offset = numpy.zeros(len(local_data_number[0]))
 
         # 8. create figure
+        bar_labels = []
         figure, axes = matplotlib.pyplot.subplots()
         for x in range(len(local_data_number)):
             if not current_needbar["horizontal"]:
@@ -343,21 +352,35 @@ def process_needbar(app: Sphinx, doctree: nodes.document, fromdocname: str):
                     color=colors[x],
                 )
 
-            if current_needbar["show_sum"]:
-                try:
-                    axes.bar_label(bar, label_type="center")  # show label in the middel of each bar
-                except AttributeError:  # bar_label is not support in older matplotlib versions
-                    current_needbar["show_sum"] = None
-
             if current_needbar["stacked"]:
+                # handle stacked bar
                 y_offset = y_offset + numpy.array(local_data_number[x])
 
-                # show for a stacked bar the overall value
-                if current_needbar["show_sum"] and x == len(local_data_number) - 1:
-                    try:
-                        axes.bar_label(bar)
-                    except AttributeError:  # bar_label is not support in older matplotlib versions
-                        current_needbar["show_sum"] = None
+            if current_needbar["show_sum"]:
+                try:
+                    bar_label = axes.bar_label(bar, label_type="center")  # show label in the middel of each bar
+                    bar_labels.append(bar_label)
+                except AttributeError:  # bar_label is not support in older matplotlib versions
+                    current_needbar["show_sum"] = None
+                    current_needbar["show_top_sum"] = None
+
+            if current_needbar["show_top_sum"] and (
+                not current_needbar["stacked"] or (x == len(local_data_number) - 1)
+            ):
+                # "show_top_sum" and ( not stacked or end of stack )
+                try:
+                    bar_label = axes.bar_label(bar)
+                    bar_labels.append(bar_label)
+                except AttributeError:  # bar_label is not support in older matplotlib versions
+                    current_needbar["show_sum"] = None
+                    current_needbar["show_top_sum"] = None
+
+        sum_rotation = current_needbar["sum_rotation"]
+        if sum_rotation and (current_needbar["show_top_sum"] or current_needbar["show_sum"]):
+            sum_rotation = sum_rotation.strip()
+            # Rotate the bar labels
+            if sum_rotation.isdigit():
+                matplotlib.pyplot.setp(bar_labels, rotation=int(sum_rotation))
 
         if not current_needbar["horizontal"]:
             # We want to support even older version of matplotlib, which do not support axes.set_xticks(labels)